+++ src/kaocha/repl.clj
@@ -79,13 +79,11 @@ These will particularly come in handy when developing plugins."}
   alternative configuration file to use with `:config-file`."
   ([]
    (config {}))
-  ([extra-config]
+  ([{:keys [profile] :as  extra-config}]
    (let [config-file (:config-file extra-config "tests.edn")
-         config      (-> (config/load-config config-file)
+         config      (-> (config/load-config (config/find-config-and-warn config-file) {:profile profile})
                          (config/merge-config (config/normalize extra-config))
-                         (cond-> #_config
-                           (.exists (io/file config-file))
-                           (assoc-in [:kaocha/cli-options :config-file] config-file)))
+                         (config/validate!))
          plugin-chain (plugin/load-all (:kaocha/plugins config))]
      (plugin/with-plugins plugin-chain
        (plugin/run-hook :kaocha.hooks/config config)))))