+++ test/unit/kaocha/watch_test.clj
@@ -1,5 +1,7 @@
 (ns kaocha.watch-test
   (:require [clojure.test :refer :all]
+            [clojure.spec.alpha :as s]
+            [kaocha.config :as c]
             [kaocha.watch :as w]
             [kaocha.platform :as platform]
             [kaocha.test-util :as util]
@@ -65,7 +67,6 @@
   (is (= "README.md" (w/convert "README.md")))
   (is (= "README.md" (w/convert "README.md "))))
 
-
 (deftest glob-converted-unchanged-test
   ; Validate that compatible patterns still match/fail to match after conversion.
   (is (w/glob? (.toPath (io/file "xxxx.clj")) [(w/convert "xxx*")]))
@@ -103,7 +104,6 @@
        (let [tmp-file (File/createTempFile "tests" ".edn")]
          (spit tmp-file "#kaocha/v1 {:tests [{:id :foo}]}")
          (first (w/reload-config {:kaocha/cli-options {:config-file (str tmp-file)}} []))))))
-
 
 (deftest ^{:min-java-version "1.11"} watch-test
   (let [{:keys [config-file test-dir] :as m} (integration/test-dir-setup {})
@@ -189,6 +189,21 @@
            @out-str))
     (is (= 0 @exit-code))))
 
+(deftest reload-test
+  (testing "reloading a configuration file produces valid config"
+    (let [orig-config (c/load-config-for-cli-and-validate "test/unit/kaocha/config/loaded-test.edn" {})
+          [reloaded-config _] (w/reload-config orig-config nil)]
+      (is (s/valid? :kaocha/config reloaded-config)
+          (s/explain :kaocha/config reloaded-config))))
+  (testing "reloading a configuration file produces the same config"
+    (let [orig-config (c/load-config-for-cli-and-validate "test/unit/kaocha/config/loaded-test.edn" {})
+          [reloaded-config _] (w/reload-config orig-config nil)]
+      (is (= orig-config reloaded-config))))
+  (testing "reloading a configuration file produces the same config when using a profile"
+    (let [orig-config (c/load-config-for-cli-and-validate "test/unit/kaocha/config/loaded-test-profile.edn" {:profile :test})
+          [reloaded-config _] (w/reload-config orig-config nil)]
+      (is (= orig-config reloaded-config)))))
+     
 ;;TODO move to cucumber
 (deftest ^{:min-java-version "1.11"} watch-load-error-test
   (let [{:keys [config-file test-dir] :as m} (integration/test-dir-setup {})